feat(flutter): rework home screen into three-pane command center
Replace separate home/schedule screens with a single Left Rail / Center Stage / Right Inspector layout driven by a shared selection provider. Panes collapse into Drawer/endDrawer overlays on narrow widths and support manual resize/drag-to-collapse independent of window size. Add ResizableTextField for drag-to-resize multi-line inputs and update related docs.

+++ PriorityTaskManager.Flutter/lib/models/task_item.dart
@@ -15,7 +15,7 @@ class TaskItem extends HiveObject {
     this.estimatedDurationMinutes = 60,
     List<String>? dependencies,
     this.importance = 5,
-    this.complexity = 1.0,
+    this.complexity = 1,
     this.notBefore,
     this.isPinned = false,
     this.isDivisible = true,
@@ -55,7 +55,7 @@ class TaskItem extends HiveObject {
   /// Cognitive load/effort, mirroring `PriorityTaskManager.Models.TaskItem.Complexity`.
   /// Not yet editable from the UI (defaults to 1.0).
   @HiveField(9)
-  double complexity;
+  int complexity;
 
   /// Earliest allowed start time, mirroring `PriorityTaskManager.Models.TaskItem.NotBefore`.
   /// Not yet editable from the UI.
@@ -81,7 +81,7 @@ class TaskItem extends HiveObject {
     int? estimatedDurationMinutes,
     List<String>? dependencies,
     int? importance,
-    double? complexity,
+    int? complexity,
     DateTime? notBefore,
     bool clearNotBefore = false,
     bool? isPinned,
